// src/app/models/user.ts
export interface User  
{
   UserID: number ;
   CompanyID: number ;
   Name: string ;
   IdentificationNum: string ;
   Sex: string ;
   Birthday: Date ;
   Picture: string ;
   IdentificationType: string ;
   Password: string ;
   EMailAddress: string ;
   CountryID: number ;
   StateID: number ;
   CityID: number ;
   Inactive: boolean ;
}